fix: The test that checks lock ordering
