+++ tests/test_unit_tests_threaded.py
@@ -3393,9 +3393,29 @@ class TestContextualBuilderWithThreads(TestBuilderWithThreads):
                     self.assertNotEqual(current_manifest["active_manifest"], thread_manifest_data[other_thread_id]["active_manifest"])
 
 
-class TestManagedResourceCrossThread(unittest.TestCase):
-    """Tests cross-thread resources handling, especially closing/releasind.
+class TestLocking(unittest.TestCase):
+    """Tests for the locks that guard native resources:
+    - the per-object operation lock that serializes native calls against teardown,
+    - the fragment lock,
+    - cross-thread creation/closing/releasing.
+
+    Every join here is bounded:
+    A deadlock must fail the test when timing out, not hang the suite.
     """
+
+    JOIN_TIMEOUT = 30
+
+    @classmethod
+    def setUpClass(cls):
+        cls.data_dir = FIXTURES_FOLDER
+        with open(DEFAULT_TEST_FILE, 'rb') as handle:
+            cls.image_bytes = handle.read()
+        with open(os.path.join(FIXTURES_FOLDER,
+                               "es256_certs.pem"), 'rb') as handle:
+            cls.certs = handle.read()
+        with open(os.path.join(FIXTURES_FOLDER,
+                               "es256_private.key"), 'rb') as handle:
+            cls.private_key = handle.read()
 
     def setUp(self):
         # Flush pending finalizers through the real free first.
@@ -3406,6 +3426,31 @@ class TestManagedResourceCrossThread(unittest.TestCase):
 
     def tearDown(self):
         ManagedResource._free_native_ptr = self._real_free
+
+    def _join_all(self, threads, what):
+        for thread in threads:
+            thread.join(self.JOIN_TIMEOUT)
+        stuck = [t for t in threads if t.is_alive()]
+        self.assertEqual(
+            stuck, [],
+            "{} did not finish within {}s: deadlock".format(
+                what, self.JOIN_TIMEOUT))
+
+    def _run_isolated(self, body, timeout=180):
+        """Run body in a subprocess and return it,
+        so that crashes can be caught and do not crash the suite itself.
+        """
+        source = textwrap.dedent(body)
+        return subprocess.run(
+            [sys.executable, "-c", source],
+            cwd=os.path.dirname(os.path.dirname(os.path.abspath(__file__))),
+            capture_output=True,
+            timeout=timeout,
+        )
+
+    def _make_signer(self):
+        return Signer.from_info(C2paSignerInfo(
+            SigningAlg.ES256, self.certs, self.private_key, None))
 
     def _free_counts(self):
         counts = {}
@@ -3509,49 +3554,6 @@ class TestManagedResourceCrossThread(unittest.TestCase):
             self.assertEqual(context_pid, pid)
             self.assertTrue(valid)
         self.assertEqual(settings._owner_pid, pid)
-
-
-class TestManagedResourceLockDeadlock(unittest.TestCase):
-    """Tests for the operation lock that serializes native calls against
-    teardown.
-
-    Every join here is bounded:
-    A deadlock must fail the test when timing out, not hang the suite.
-    """
-
-    JOIN_TIMEOUT = 30
-
-    @classmethod
-    def setUpClass(cls):
-        with open(DEFAULT_TEST_FILE, 'rb') as handle:
-            cls.image_bytes = handle.read()
-        with open(os.path.join(FIXTURES_FOLDER,
-                               "es256_certs.pem"), 'rb') as handle:
-            cls.certs = handle.read()
-        with open(os.path.join(FIXTURES_FOLDER,
-                               "es256_private.key"), 'rb') as handle:
-            cls.private_key = handle.read()
-
-    def _join_all(self, threads, what):
-        for thread in threads:
-            thread.join(self.JOIN_TIMEOUT)
-        stuck = [t for t in threads if t.is_alive()]
-        self.assertEqual(
-            stuck, [],
-            "{} did not finish within {}s: deadlock".format(
-                what, self.JOIN_TIMEOUT))
-
-    def _run_isolated(self, body, timeout=180):
-        """Run body in a subprocess and return it,
-        so that crashes can be caught and do not crash the suite itself.
-        """
-        source = textwrap.dedent(body)
-        return subprocess.run(
-            [sys.executable, "-c", source],
-            cwd=os.path.dirname(os.path.dirname(os.path.abspath(__file__))),
-            capture_output=True,
-            timeout=timeout,
-        )
 
     def test_json_racing_finalizer_does_not_crash(self):
         """Readers used on one thread while others are collected.
@@ -4355,148 +4357,18 @@ class TestManagedResourceLockDeadlock(unittest.TestCase):
             "borrowed handles used without their own guard:\n  "
             + "\n  ".join(unguarded))
 
-    def test_no_conflicting_lock_acquisition_order(self):
-        """No two locks may be nested in opposite orders by different methods.
-
-        Two methods nesting the same pair of locks in opposite order is a
-        AB/BA deadlock shape: thread 1 holds A and waits for B while
-        thread 2 holds B and waits for A.
-
-        This scans the code to check.
-        """
-        module = sys.modules[Reader.__module__]
-        tree = ast.parse(inspect.getsource(module))
-
-        # Every self._X = threading.Lock()/RLock()/Condition() assignment,
-        # grouped by the class that owns it.
-        lock_attrs_by_class = {}
-        for cls in ast.walk(tree):
-            if not isinstance(cls, ast.ClassDef):
-                continue
-            found = set()
-            for node in ast.walk(cls):
-                if not (isinstance(node, ast.Assign)
-                        and len(node.targets) == 1):
-                    continue
-                target = node.targets[0]
-                if not (isinstance(target, ast.Attribute)
-                        and isinstance(target.value, ast.Name)
-                        and target.value.id == "self"):
-                    continue
-                value = node.value
-                if (isinstance(value, ast.Call)
-                        and isinstance(value.func, ast.Attribute)
-                        and value.func.attr in
-                        ("Lock", "RLock", "Condition")):
-                    found.add(target.attr)
-            if found:
-                lock_attrs_by_class[cls.name] = found
-
-        def lock_name_for_with(item):
-            """The lock attribute a `with` item enters, or None."""
-            ctx = item.context_expr
-            # with self._fragment_lock:
-            if (isinstance(ctx, ast.Attribute)
-                    and isinstance(ctx.value, ast.Name)
-                    and ctx.value.id == "self"
-                    and any(ctx.attr in attrs
-                            for attrs in lock_attrs_by_class.values())):
-                return ctx.attr
-            # with self._lock(): returns _op_lock itself.
-            if (isinstance(ctx, ast.Call)
-                    and isinstance(ctx.func, ast.Attribute)
-                    and ctx.func.attr == "_lock"
-                    and isinstance(ctx.func.value, ast.Name)
-                    and ctx.func.value.id == "self"):
-                return "_op_lock"
-            return None
-
-        def orders_in(node, stack, pairs):
-            """Record (outer, inner) for every nesting this node contains."""
-            if isinstance(node, (ast.With, ast.AsyncWith)):
-                names = [n for n in (lock_name_for_with(i)
-                                     for i in node.items) if n]
-                for name in names:
-                    if stack:
-                        pairs.add((stack[-1], name))
-                    stack.append(name)
-                for child in node.body:
-                    orders_in(child, stack, pairs)
-                for _ in names:
-                    stack.pop()
-                return
-            for child in ast.iter_child_nodes(node):
-                orders_in(child, stack, pairs)
-
-        pairs_by_method = {}
-        for cls in ast.walk(tree):
-            if not isinstance(cls, ast.ClassDef):
-                continue
-            for method in cls.body:
-                if not isinstance(method, (ast.FunctionDef,
-                                           ast.AsyncFunctionDef)):
-                    continue
-                pairs = set()
-                orders_in(method, [], pairs)
-                if pairs:
-                    pairs_by_method[(cls.name, method.name)] = pairs
-
-        all_pairs = set().union(*pairs_by_method.values()) \
-            if pairs_by_method else set()
-        conflicts = []
-        for (outer, inner) in all_pairs:
-            # Each conflicting pair appears twice in all_pairs, once per direction.
-            if outer >= inner or (inner, outer) not in all_pairs:
-                continue
-            forward = [k for k, v in pairs_by_method.items()
-                      if (outer, inner) in v]
-            backward = [k for k, v in pairs_by_method.items()
-                       if (inner, outer) in v]
-            conflicts.append(
-                "{} nests {} inside {}, but {} nests {} inside {}".format(
-                    forward[0], inner, outer, backward[0], outer, inner))
-
-        self.assertGreater(
-            len(pairs_by_method), 0,
-            "lock nesting scan found no nested lock acquisitions: "
-            "the scan is broken")
-        self.assertEqual(
-            conflicts, [],
-            "conflicting lock acquisition order:\n  "
-            + "\n  ".join(sorted(set(conflicts))))
-
-
-class TestSharedSignerTeardownRace(unittest.TestCase):
-    """A Signer shared across threads must not be freed mid-sign.
-
-    Builder.sign borrows the signer's handle for the duration of the native
-    call. Without a guard on the signer itself, a close() on another thread
-    frees that handle while c2pa_builder_sign is using it, and the process
-    dies with SIGSEGV instead of raising.
-    """
-
-    def setUp(self):
-        self.data_dir = os.path.join(os.path.dirname(__file__), "fixtures")
-        with open(os.path.join(self.data_dir, "C.jpg"), "rb") as f:
-            self.image_bytes = f.read()
-        with open(os.path.join(self.data_dir, "es256_certs.pem"), "rb") as f:
-            self.certs = f.read()
-        with open(os.path.join(self.data_dir, "es256_private.key"), "rb") as f:
-            self.key = f.read()
-        self.manifest = {
-            "claim_generator_info": [{"name": "test", "version": "0.1"}],
-            "assertions": [],
-        }
-
-    def _make_signer(self):
-        return Signer.from_info(C2paSignerInfo(
-            SigningAlg.ES256, self.certs, self.key, None))
 
     def test_close_during_concurrent_sign_does_not_crash(self):
-        """Rotate a shared signer while other threads sign with it.
+        """A Signer shared across threads must not be freed mid-sign.
 
-        Runs in a subprocess: the failure mode is a segfault, which would
-        take the test runner down with it otherwise.
+        Builder.sign borrows the signer's handle for the duration of the
+        native call. Without a guard on the signer itself, a close() on
+        another thread frees that handle while c2pa_builder_sign is using
+        it, and the process dies with SIGSEGV instead of raising.
+
+        Rotates a shared signer while other threads sign with it. Runs in a
+        subprocess: the failure mode is a segfault, which would take the
+        test runner down with it otherwise.
         """
         source = textwrap.dedent("""
             import io, os, sys, threading
